+++ syseng_hwco/proj/Artifact/Code/SystemC_TLM/include/Audio/DACSim.h
@@ -8,11 +8,11 @@
 class DACSim : public sc_module
 {
 public:
-  sc_fifo_in<int> data_from_splitter;
+  sc_fifo_in<short> data_from_splitter;
 
   // This is to simulate that the data from the Speakers are
   // also picked up by the microphone
-  sc_fifo_out<int> data_to_adcsim;
+  sc_fifo_out<short> data_to_adcsim;
 
 private:
   void dac_writer_thread();
