+++ syseng_hwco/proj/Artifact/Code/SystemC_TLM/include/Audio/EchoCancellation.h
@@ -10,10 +10,10 @@
 class EchoCancellation : public sc_module
 {
 public:
-  sc_fifo_in<int> data_from_adc;
-  sc_fifo_in<int> data_from_splitter;
+  sc_fifo_in<short> data_from_adc;
+  sc_fifo_in<short> data_from_splitter;
 
-  sc_fifo_out<int> data_to_audio_encoding;
+  sc_fifo_out<short> data_to_audio_encoding;
 
 private:
   void echo_cancellation_thread();
