+++ syseng_hwco/proj/Artifact/Code/SystemC_TLM/include/Audio/Splitter.h
@@ -6,10 +6,10 @@
 class Splitter : public sc_module
 {
 public:
-  sc_fifo_in<int> data_from_decoding;
+  sc_fifo_in<short> data_from_decoding;
 
-  sc_fifo_out<int> data_to_dac;
-  sc_fifo_out<int> data_to_echocancellation;
+  sc_fifo_out<short> data_to_dac;
+  sc_fifo_out<short> data_to_echocancellation;
 
 private:
   void splitter_thread();
